% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demoCatchment}
\alias{demoCatchment}
\title{Demonstration project}
\usage{
demoCatchment(cs_dir = tempdir(TRUE))
}
\arguments{
\item{cs_dir}{An optional character string specifying an existing directory.}
}
\value{
A character string containing the demonstration project root
directory.
}
\description{
Copies a demonstration project to an existing or a temporary directory.
}
