% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demoProject}
\alias{demoProject}
\title{Demonstration project}
\usage{
demoProject(cs_dir = tempdir(TRUE))
}
\arguments{
\item{cs_dir}{An optional character string specifying an existing directory.}
}
\value{
A character string containing the demonstration project root
directory.
}
\description{
Copies a demonstration project to an existing or a temporary directory.
}
