% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{DEMrelatedInput}
\alias{DEMrelatedInput}
\title{DEM Related Input}
\usage{
DEMrelatedInput(cv_dir, cs_dem, cs_cha, sp_msk, sp_olp, sp_sds,
  cs_rds = NULL, cs_wgs = NULL, ns_brn = 50, is_adj = 1L,
  is_ths = 1L, ls_tmp = FALSE)
}
\arguments{
\item{cv_dir}{A character vector specifying the desired project root
directory (first position).}

\item{cs_dem}{A character string specifying a path to a potentially large
raster digital elevation model.}

\item{cs_cha}{A character string specifying a path to a potentially large
raster providing channels.}

\item{sp_msk}{An \code{\link[sp:SpatialPolygonsDataFrame-class]{sp::SpatialPolygonsDataFrame}} providing a somewhat
oversized catchment mask used to clip the potentially large input rasters
for further processing.}

\item{sp_olp}{An \code{\link[sp:SpatialPointsDataFrame-class]{sp::SpatialPointsDataFrame}} providing the desired
catchment outlet.}

\item{sp_sds}{An \code{\link[sp:SpatialPointsDataFrame-class]{sp::SpatialPointsDataFrame}} providing the mapped
channel sources.}

\item{cs_rds}{An optional character string specifying a path to a potentially
large raster providing roads.}

\item{cs_wgs}{An optional character string specifying a path to a potentially
large raster providing flow accumulation weights.}

\item{ns_brn}{A numeric scalar specifying the stream burning step size in m.}

\item{is_adj}{A numeric scalar specifying how many cells adjacent to channels
shall be burnt.}

\item{is_ths}{An integer scalar specifying the number of threads to use
during computation.}

\item{ls_tmp}{A logical scalar specifying if the temporary files created
during computation shall be kept.}
}
\value{
A numeric \code{\link{matrix}} specifying the catchment outlet coordinates.
}
\description{
Clips and calculates or derives all input data related to the digital
elevation model in the broader sense: \emph{acc, acc_wtd, cha, dem, dir, rds,
slp,} and \emph{wsh.}

Requires \href{http://hydrology.usu.edu/taudem/taudem5/downloads.html}{TauDEM}
5.3.7 to be installed on your computer.
}
\details{
In case no flow accumulation weights are provided, \emph{acc} and \emph{acc_wtd}
are identical.

\emph{slp} represents D8 slopes.
}
\references{
\cite{Lindsay, J.B., Creed, I.F., 2005. Removal of artifact depressions from
digital elevation models: towards a minimum impact approach. Hydrol. Process.
19, 3113–3126.}
}
\seealso{
\code{\link{RPhosFate}}, \code{\link{catchment}}
}
