% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{img2tif}
\alias{img2tif}
\title{Convert \emph{ERDAS IMAGINE} to \emph{GeoTIFF} raster files}
\usage{
img2tif(cs_dir, cs_crs = NULL)
}
\arguments{
\item{cs_dir}{A character string specifying an existing directory.}

\item{cs_crs}{An optional character string used to set the coordinate
reference system of all output raster files. See \code{\link[terra:crs]{terra::crs}} for further
information.}
}
\value{
A character vector containing the paths to the processed \emph{ERDAS
IMAGINE} raster files.
}
\description{
Converts all \emph{ERDAS IMAGINE} raster files in a directory and its
subdirectories into \emph{GeoTIFF} raster files.
}
