% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{saveState,RPhosFate-method}
\alias{saveState,RPhosFate-method}
\alias{saveState}
\title{Save state}
\usage{
\S4method{saveState}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\link[=RPhosFate-class]{RPhosFate}} river catchment object.}
}
\value{
\code{NULL} invisibly and side effects in the form of files.
}
\description{
Saves parameters \emph{(parameters.yaml)} to disk.
}
\examples{

# temporary demonstration project copy
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)

saveState(x)

}
\seealso{
\code{\link{RPhosFate}}, \code{\link{catchment}}
}
