% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{transportCalcOrder,RPhosFate-method}
\alias{transportCalcOrder,RPhosFate-method}
\alias{transportCalcOrder}
\title{Transport Calculation Order}
\usage{
\S4method{transportCalcOrder}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Determines the cell transport calculation order.
}
