% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{autoCalibrate,RPhosFate-method}
\alias{autoCalibrate,RPhosFate-method}
\alias{autoCalibrate}
\title{Automatic model calibration}
\usage{
\S4method{autoCalibrate}{RPhosFate}(x, substance, col, interval, metric,
  tol = min(interval) * 0.1, parameter = NULL)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}

\item{col}{A character string specifying the calibration data column with the
respective substance loads.}

\item{interval}{A numeric vector specifying the end-points of the interval to
be searched.}

\item{metric}{A character string specifying the metric to optimise. See
\code{\link{calibrationQuality}} for available metrics.}

\item{tol}{A numeric scalar specifying the desired accuracy of the parameter
used for optimisation (not the metric).}

\item{parameter}{By default, SS are optimised utilising the overland
deposition rate and all other substances are optimised utilising their
respective enrichment ratio. This argument can be used to specify a
dedicated parameter utilised for optimisation via a character string
(overland (\code{"ns_dep_ovl"}) or channel deposition rate (\code{"ns_dep_cha"})).}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Automatically calibrates the model with the help of a combination of golden
section search and successive parabolic interpolation.
}
\seealso{
\code{\link{optimize}}
}
