% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{autoCalibrate2,RPhosFate-method}
\alias{autoCalibrate2,RPhosFate-method}
\alias{autoCalibrate2}
\title{Multidimensional automatic model calibration}
\usage{
\S4method{autoCalibrate2}{RPhosFate}(
  x,
  substance,
  col,
  interval,
  metric,
  method = c("Nelder-Mead", "SANN"),
  control = list(fnscale = if (metric \%in\% c("NSE", "mNSE")) -1 else 1)
)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}

\item{col}{A character string specifying the calibration data column with the
respective substance river loads.}

\item{interval}{A numeric vector specifying the end-points of the interval to
be searched.}

\item{metric}{A character string specifying the metric to optimise. See
\code{\link{calibrationQuality}} for available metrics.}

\item{method}{A character string specifying the utilised optimisation method.
Either \code{"Nelder-Mead"} or \code{"SANN"} for simulated annealing.}

\item{control}{A \code{\link{list}} of control parameters passed on to \code{\link{optim}}.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object and side effects in
the form of raster files.
}
\description{
Automatically calibrates the model with the help of the Nelder–Mead or a
simulated annealing optimisation method. In contrast to \code{\link{autoCalibrate}},
this method calibrates the overland and channel deposition rate in one go.
Beware of local optima.
}
\examples{
\dontrun{
# temporary demonstration project copy
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)
# presupposed method calls
x <- firstRun(x, "SS")
x <- snapGauges(x)

x <- autoCalibrate2(
  x,
  "SS",
  col = "SS_load",
  interval = c(1e-6, 1e-2),
  metric = "NSE",
  control = list(fnscale = -1, parscale = c(1e-3, 1e-3))
)
}

}
\seealso{
\code{\link{snapGauges}}
}
