% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{emission,RPhosFate-method}
\alias{emission,RPhosFate-method}
\alias{emission}
\title{Emission}
\usage{
\S4method{emission}{RPhosFate}(cmt, substance = "PP")
}
\arguments{
\item{cmt}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates substance emissions.
}
