% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{erosion,RPhosFate-method}
\alias{erosion,RPhosFate-method}
\alias{erosion}
\title{Erosion}
\usage{
\S4method{erosion}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates (R)USLE erosion.
}
\references{
\cite{Renard, K.G., Foster, G.R., Weesies, G.A., McCool, D.K., Yoder, D.C.,
1997. Predicting soil erosion by water: a guide to conservation planning with
the Revised Universal Soil Loss Equation (RUSLE), Agriculture Handbook. U.S.
Government Printing Office, Washington, DC.}

\cite{Wischmeier, W.H., Smith, D.D., 1978. Predicting rainfall erosion
losses. A guide to conservation planning, Agriculture Handbook. U.S.
Government Printing Office, Washington, DC.}
}
