% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{erosionPrerequisites,RPhosFate-method}
\alias{erosionPrerequisites,RPhosFate-method}
\alias{erosionPrerequisites}
\title{Erosion prerequisites}
\usage{
\S4method{erosionPrerequisites}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates capped slopes, L- and RUSLE S-factors. Weighted flow accumulations
less than one are set to one for the calculation of the L-factors.
}
\examples{
\dontrun{
# create temporary demonstration project
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)

x <- erosionPrerequisites(x)
}

}
\references{
\cite{Desmet, P.J.J., Govers, G., 1996. A GIS procedure for automatically
calculating the USLE LS factor on topographically complex landscape units.
Journal of Soil and Water Conservation 51, 427–433.}

\cite{Renard, K.G., Foster, G.R., Weesies, G.A., McCool, D.K., Yoder, D.C.,
1997. Predicting soil erosion by water: a guide to conservation planning with
the Revised Universal Soil Loss Equation (RUSLE), Agriculture Handbook. U.S.
Government Printing Office, Washington, DC.}
}
