% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{firstRun,RPhosFate-method}
\alias{firstRun,RPhosFate-method}
\alias{firstRun}
\title{First Run}
\usage{
\S4method{firstRun}{RPhosFate}(x, substance = "PP")
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calls \code{\link{erosionPrerequisites}}, \code{\link{erosion}}, \code{\link{emission}},
\code{\link{transportPrerequisites}}, \code{\link{transportCalcOrder}} and \code{\link{transport}} in the
mentioned order. While \code{\link{transport}} is called for the specified substance
only, \code{\link{emission}} is called for all substances whose top soil concentrations
have been provided.
}
\seealso{
\code{\link{subsequentRun}}
}
