% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettersSetters.R
\docType{methods}
\name{getLayer,RPhosFate-method}
\alias{getLayer,RPhosFate-method}
\alias{getLayer}
\alias{[,RPhosFate,ANY,ANY,ANY-method}
\title{Get Layer}
\usage{
\S4method{getLayer}{RPhosFate}(x, i, j = NULL)

\S4method{[}{RPhosFate,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{i}{A character string specifying a layer name. Substance related layers
whose names start with \emph{xx} are treated differently. They have to be
queried by their name (not filename), for example, \code{"xxc"} together with
the respective substance specified in argument \code{j}, for example, \code{"PP"},
which queries the particulate phosphorus concentration in top soil. See
subdirectory sections for further information.}

\item{j}{A character string specifying a substance if applicable.}
}
\value{
A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}} object.
}
\description{
Obtains a project raster layer for further analysis.
}
\section{\emph{Input} subdirectory}{

This directory holds all possible user input raster data:
\itemize{
\item \emph{acc:} Flow accumulations required for transport (the top-most cells must
have a value of one).
\item \emph{acc_wtd:} Weighted flow accumulations (can be equal to \emph{acc}).
\item \emph{CFa:} (R)USLE C-factors.
\item \emph{cha:} Channel cells (1: channel cell, \code{NA}: no channel cell).
\item \emph{clc:} Clay content of top soils in \% required for substance emissions.
\item \emph{dem:} Digital elevation model in m a.s.l. (optional).
\item \emph{dir:} D8 flow directions required for transport.
\item \emph{fid:} Field IDs (optional).
\item \emph{KFa:} (R)USLE K-factors.
\item \emph{lue:} Land use classes (optional).
\item \emph{man:} Manning's roughness coefficients required for transport.
\item \emph{xxc:} Substance content of top soils in mg/kg required for substance
emissions, for example, \emph{ppc} (PP).
\item \emph{rds:} Road cells required for transport (0: road cell without subsurface
drainage, 1: road cell with subsurface drainage, \code{NA}: no road cell).
\item \emph{RFa:} (R)USLE R-factors.
\item \emph{slp:} Slopes in \%.
\item \emph{wsh:} Watershed (optional).
}
}

\section{\emph{Intermediate} subdirectory}{

This directory holds intermediate results:
\itemize{
\item \emph{inl:} Cells representing inlets at roads (storm drains).
\item \emph{LFa:} L-factors.
\item \emph{rhy:} Hydraulic radii in m.
\item \emph{rip:} Cells representing the riparian zones within channel cells.
\item \emph{SFa:} RUSLE S-factors.
\item \emph{slp_cap:} Capped slopes in \%.
}
}

\section{\emph{Result} subdirectory}{

This directory holds the model results:
\itemize{
\item \emph{ero:} Erosion in t/cell/yr.
\item \emph{xxe:} Substance emissions in kg/cell/yr, for example, \emph{ppe} (PP).
\item \emph{xxr:} Substance retentions in t/cell/yr (SS) or kg/cell/yr, for example,
\emph{ppr} (PP).
\item \emph{xxt:} Substance transports in t/cell/yr (SS) or kg/cell/yr, for example,
\emph{ppt} (PP).
\item \emph{xxt_cld:} Substance cell loads in t/cell/yr (SS) or kg/cell/yr, for
example, \emph{ppt_cld} (PP).
\item \emph{xxt_ctf:} Substance cell transfers in t/cell/yr (SS) or kg/cell/yr,
for example, \emph{ppt_ctf} (PP).
\item \emph{xxt_inp:} Substance inputs into surface waters in t/cell/yr (SS) or
kg/cell/yr, for example, \emph{ppt_inp} (PP).
\item \emph{xxt_out:} Substance outlet loads in t/cell/yr (SS) or kg/cell/yr, for
example, \emph{ppt_out} (PP).
}
}

