% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{img2tif}
\alias{img2tif}
\title{Convert ERDAS IMAGINE to GeoTIFF raster files}
\usage{
img2tif(cs_dir)
}
\arguments{
\item{cs_dir}{A character string specifying an existing directory.}
}
\value{
A character vector containing the paths to the converted ERDAS
IMAGINE raster files.
}
\description{
Converts all ERDAS IMAGINE raster files in a directory and its subdirectories
into GeoTIFF raster files.
}
