% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{saveState,RPhosFate-method}
\alias{saveState,RPhosFate-method}
\alias{saveState}
\title{Save state}
\usage{
\S4method{saveState}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
\code{NULL} invisibly.
}
\description{
Saves parameters \emph{(parameters.yaml)} and transport calculation order
\emph{(order.rds)} to disk.
}
