% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{snapGauges,RPhosFate-method}
\alias{snapGauges,RPhosFate-method}
\alias{snapGauges}
\title{Snap Gauge(s)}
\usage{
\S4method{snapGauges}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Snaps the coordinates of all calibration gauges to the midpoint of the
nearest channel cell.
}
\seealso{
\code{\link{calibrationQuality}}
}
