% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{subsequentRun,RPhosFate-method}
\alias{subsequentRun,RPhosFate-method}
\alias{subsequentRun}
\title{Subsequent run}
\usage{
\S4method{subsequentRun}{RPhosFate}(x, substance = "PP")
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calls \code{\link{transport}} for the specified substance. In Monte Carlo simulation
mode \code{\link{erosion}} and, if applicable, \code{\link{emission}} are called as well.
}
\examples{
\dontrun{
# create temporary demonstration project
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)
# presupposed function call
x <- firstRun(x, "SS")

x <- subsequentRun(x, "PP")
}

}
\seealso{
\code{\link{firstRun}}
}
