% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{subsequentRun,RPhosFate-method}
\alias{subsequentRun,RPhosFate-method}
\alias{subsequentRun}
\title{Subsequent Run}
\usage{
\S4method{subsequentRun}{RPhosFate}(x, substance = "PP")
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calls \code{\link{transport}} for the specified substance. In Monte Carlo simulation
mode \code{\link{erosion}} and, if applicable, \code{\link{emission}} are called as well.
}
\seealso{
\code{\link{firstRun}}
}
