% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{transport,RPhosFate-method}
\alias{transport,RPhosFate-method}
\alias{transport}
\title{Transport}
\usage{
\S4method{transport}{RPhosFate}(cmt, substance = "PP")
}
\arguments{
\item{cmt}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}

\item{substance}{A character string specifying the substance to calculate.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates substance retentions, transports and cell loads as well as
transfers among others.
}
\references{
\cite{Engman, E.T., 1986. Roughness coefficients for routing surface runoff.
Journal of Irrigation and Drainage Engineering 112, 39–53.}
}
