% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\name{transportCalcOrder,RPhosFate-method}
\alias{transportCalcOrder,RPhosFate-method}
\alias{transportCalcOrder}
\title{Transport calculation order}
\usage{
\S4method{transportCalcOrder}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Determines the cell transport calculation order.
}
\examples{
\dontrun{
# create temporary demonstration project
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)

x <- transportCalcOrder(x)
}

}
\seealso{
\code{\link{firstRun}}, \code{\link{subsequentRun}}
}
