% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{transportPrerequisites,RPhosFate-method}
\alias{transportPrerequisites,RPhosFate-method}
\alias{transportPrerequisites}
\title{Transport Prerequisites}
\usage{
\S4method{transportPrerequisites}{RPhosFate}(cmt)
}
\arguments{
\item{cmt}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates hydraulic radii and determines cells representing inlets as well
as riparian zones.
}
