% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhosFate.R
\docType{methods}
\name{transportPrerequisites,RPhosFate-method}
\alias{transportPrerequisites,RPhosFate-method}
\alias{transportPrerequisites}
\title{Transport Prerequisites}
\usage{
\S4method{transportPrerequisites}{RPhosFate}(x)
}
\arguments{
\item{x}{An S4 \code{\linkS4class{RPhosFate}} river catchment object.}
}
\value{
An S4 \code{\linkS4class{RPhosFate}} river catchment object.
}
\description{
Calculates hydraulic radii and determines cells representing inlets as well
as riparian zones.
}
\references{
\cite{Molnár, P., Ramírez, J.A., 1998. Energy dissipation theories and
optimal channel characteristics of river networks. Water Resources Research
34, 1809–1818.}
}
